requirements.txt

Django
djangorestframework
celery[redis]
redis
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests
beautifulsoup4
lxml
python-dotenv
ipython
